requirements.txt
aiohappyeyeballs==2.4.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
certifi==2025.1.31
cffi==1.17.1
contextlib2==21.6.0
cryptography==44.0.0
Django==5.1.4
django-ckeditor-5==0.2.15
django-cors-headers==4.7.0
django-easy-audit==1.3.7
django-extensions==3.2.3
django-js-asset==3.0.1
django-ninja==1.3.0
django-ninja-extra==0.22.3
django-ninja-jwt==5.3.5
django-redis==5.4.0
dnspython==2.7.0
email_validator==2.2.0
frozenlist==1.5.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
injector==0.22.0
multidict==6.1.0
pillow==11.1.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
qrcode==8.0
redis==5.2.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
yarl==1.18.3
